introduce CollectionViewRotationHandling protocol
