+++ novawallet/Modules/NovaMainAppContainer/NovaMainAppContainerViewController.swift
@@ -1,7 +1,9 @@
 import UIKit
 import SnapKit
 
-final class NovaMainAppContainerViewController: UIViewController, ViewHolder {
+final class NovaMainAppContainerViewController: UIViewController,
+    ViewHolder,
+    CollectionViewRotationHandling {
     typealias RootViewType = NovaMainAppContainerViewLayout
 
     let presenter: NovaMainAppContainerPresenterProtocol
@@ -133,36 +135,6 @@ private extension NovaMainAppContainerViewController {
                 return self?.rootView
             }
         )
-    }
-
-    func findCollectionView(in view: UIView) -> UICollectionView? {
-        // BFS
-        var queue = [UIView]()
-        queue.append(view)
-
-        while !queue.isEmpty {
-            let view = queue.removeFirst()
-
-            if let collectionView = view as? UICollectionView {
-                return collectionView
-            }
-
-            queue.append(contentsOf: view.subviews)
-        }
-
-        return nil
-    }
-
-    func updateCollectionViewLayoutIfNeeded() {
-        guard
-            let topView = tabBar?.topViewController()?.view,
-            let collectionView = findCollectionView(in: topView)
-        else { return }
-
-        UIView.performWithoutAnimation {
-            collectionView.collectionViewLayout.invalidateLayout()
-            collectionView.layoutIfNeeded()
-        }
     }
 
     func updateModalsLayoutIfNeeded() {
